Split Route file
